Bug caused by unusable fuel in FG 2019.1 RC
Engines did not start in FlightGear nightly begining with 2019.1. Reason was the "<ususable-gal_us>" tag of our feeder tanks. Our tank size is 0.5 "lbs" and the unusable amount is 0.1 "US-gallons" (which is 6.6 lbs - so for the engine our tank was empty). This feature was not supported by JSBSim before. https://github.com/JSBSim-Team/jsbsim/issues/141